requirements.txt
# Google ADK ve AI
google-adk
google-cloud-aiplatform
google-generativeai
vertexai

# Google APIs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Microsoft Graph
msal
requests

# Email
icalendar

# Async support
asyncio

# Environment
python-dotenv

# Utilities
python-dateutil